Refactor file structure and update imports
Renamed App.jsx to index.jsx and app.global.css to index.css for consistency. Updated import paths in main.jsx to reflect these changes.
